Soft migration of sys -> util, Removal of deprecated utils module.
