+++ lib/assert.js
@@ -23,7 +23,7 @@
 // WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 // UTILITY
-var inherits = require('sys').inherits;
+var util = require('util');
 var pSlice = Array.prototype.slice;
 
 // 1. The assert module provides functions that throw
@@ -47,7 +47,7 @@ assert.AssertionError = function AssertionError (options) {
     Error.captureStackTrace(this, stackStartFunction);
   }
 };
-inherits(assert.AssertionError, Error);
+util.inherits(assert.AssertionError, Error);
 
 assert.AssertionError.prototype.toString = function() {
   if (this.message) {
