+++ lib/child_process.js
@@ -1,4 +1,4 @@
-var inherits = require('sys').inherits;
+var util = require('util');
 var EventEmitter = require('events').EventEmitter;
 var Stream = require('net').Stream;
 var InternalChildProcess = process.binding('child_process').ChildProcess;
@@ -155,7 +155,7 @@ function ChildProcess () {
 
   this.__defineGetter__('pid', function () { return internal.pid; });
 }
-inherits(ChildProcess, EventEmitter);
+util.inherits(ChildProcess, EventEmitter);
 
 
 ChildProcess.prototype.kill = function (sig) {
