+++ lib/crypto.js
@@ -3582,7 +3582,6 @@ var RootCaCerts = [
 +"-----END CERTIFICATE-----\n"
 ];
 
-var sys = require("sys");
 try {
   var binding = process.binding('crypto');
   var SecureContext = binding.SecureContext;
