+++ lib/dgram.js
@@ -1,4 +1,4 @@
-var sys = require("sys");
+var util = require("util");
 var fs = require("fs");
 var events = require("events");
 var dns = require('dns');
@@ -72,7 +72,7 @@ function Socket (type, listener) {
   }
 }
 
-sys.inherits(Socket, events.EventEmitter);
+util.inherits(Socket, events.EventEmitter);
 exports.Socket = Socket;
 
 exports.createSocket = function (type, listener) {
