+++ lib/fs.js
@@ -1,4 +1,4 @@
-var sys = require('sys');
+var util = require('util');
 var events = require('events');
 
 var binding = process.binding('fs');
@@ -666,7 +666,7 @@ var ReadStream = fs.ReadStream = function(path, options) {
     self._read();
   });
 };
-sys.inherits(ReadStream, events.EventEmitter);
+util.inherits(ReadStream, events.EventEmitter);
 
 fs.FileReadStream = fs.ReadStream; // support the legacy name
 
@@ -761,7 +761,7 @@ var readStreamForceCloseWarning;
 ReadStream.prototype.forceClose = function (cb) {
   if (!readStreamForceCloseWarning) {
     readStreamForceCloseWarning = "ReadStream.prototype.forceClose renamed to destroy()";
-    sys.error(readStreamForceCloseWarning);
+    util.error(readStreamForceCloseWarning);
   }
   return this.destroy(cb);
 };
@@ -844,7 +844,7 @@ var WriteStream = fs.WriteStream = function(path, options) {
     this.flush();
   }
 };
-sys.inherits(WriteStream, events.EventEmitter);
+util.inherits(WriteStream, events.EventEmitter);
 
 fs.FileWriteStream = fs.WriteStream; // support the legacy name
 
@@ -940,7 +940,7 @@ var writeStreamCloseWarning;
 WriteStream.prototype.close = function (cb) {
   if (!writeStreamCloseWarning) {
     writeStreamCloseWarning = "WriteStream.prototype.close renamed to end()";
-    sys.error(writeStreamCloseWarning);
+    util.error(writeStreamCloseWarning);
   }
   return this.end(cb);
 };
@@ -958,7 +958,7 @@ var writeStreamForceCloseWarning;
 WriteStream.prototype.forceClose = function (cb) {
   if (!writeStreamForceCloseWarning) {
     writeStreamForceCloseWarning = "WriteStream.prototype.forceClose renamed to destroy()";
-    sys.error(writeStreamForceCloseWarning);
+    util.error(writeStreamForceCloseWarning);
   }
   return this.destroy(cb);
 };
