+++ lib/http.js
@@ -1,9 +1,9 @@
-var sys = require('sys');
+var util = require('util');
 
 var debug;
 var debugLevel = parseInt(process.env.NODE_DEBUG, 16);
 if (debugLevel & 0x4) {
-  debug = function (x) { sys.error('HTTP: ' + x); };
+  debug = function (x) { util.error('HTTP: ' + x); };
 } else {
   debug = function () { };
 }
@@ -207,7 +207,7 @@ function IncomingMessage (socket) {
   this.statusCode = null;
   this.client = this.socket;
 }
-sys.inherits(IncomingMessage, stream.Stream);
+util.inherits(IncomingMessage, stream.Stream);
 exports.IncomingMessage = IncomingMessage;
 
 IncomingMessage.prototype._parseQueryString = function () {
@@ -305,7 +305,7 @@ function OutgoingMessage (socket) {
 
   this.finished = false;
 }
-sys.inherits(OutgoingMessage, stream.Stream);
+util.inherits(OutgoingMessage, stream.Stream);
 exports.OutgoingMessage = OutgoingMessage;
 
 // This abstract either writing directly to the socket or buffering it.
@@ -499,7 +499,7 @@ OutgoingMessage.prototype.write = function (chunk, encoding) {
     if (typeof(chunk) === 'string') {
       len = Buffer.byteLength(chunk, encoding);
       var chunk = len.toString(16) + CRLF + chunk + CRLF;
-      debug('string chunk = ' + sys.inspect(chunk));
+      debug('string chunk = ' + util.inspect(chunk));
       ret = this._send(chunk, encoding);
     } else {
       // buffer
@@ -605,7 +605,7 @@ function ServerResponse (req) {
     this.shouldKeepAlive = false;
   }
 }
-sys.inherits(ServerResponse, OutgoingMessage);
+util.inherits(ServerResponse, OutgoingMessage);
 exports.ServerResponse = ServerResponse;
 
 ServerResponse.prototype.writeContinue = function () {
@@ -678,7 +678,7 @@ function ClientRequest (socket, method, url, headers) {
 
   this._storeHeader(method + " " + url + " HTTP/1.1\r\n", headers);
 }
-sys.inherits(ClientRequest, OutgoingMessage);
+util.inherits(ClientRequest, OutgoingMessage);
 exports.ClientRequest = ClientRequest;
 
 ClientRequest.prototype.finish = function () {
@@ -769,7 +769,7 @@ function Server (requestListener) {
 
   this.addListener("connection", connectionListener);
 }
-sys.inherits(Server, net.Server);
+util.inherits(Server, net.Server);
 
 Server.prototype.setSecure = function (credentials) {
   this.secure = true;
@@ -934,14 +934,14 @@ function Client ( ) {
       this.setSecure(this.credentials);
     } else {
       self._initParser();
-      debug('requests: ' + sys.inspect(self._outgoing));
+      debug('requests: ' + util.inspect(self._outgoing));
       outgoingFlush(self);
     }
   });
 
   self.addListener("secure", function () {
     self._initParser();
-    debug('requests: ' + sys.inspect(self._outgoing));
+    debug('requests: ' + util.inspect(self._outgoing));
     outgoingFlush(self);
   });
 
@@ -968,7 +968,7 @@ function Client ( ) {
     }
   });
 };
-sys.inherits(Client, net.Stream);
+util.inherits(Client, net.Stream);
 
 exports.Client = Client;
