+++ lib/net.js
@@ -1,4 +1,4 @@
-var sys = require("sys");
+var util = require("util");
 var fs = require("fs");
 var events = require("events");
 var stream = require("stream");
@@ -9,7 +9,7 @@ var kPoolSize = 40*1024;
 
 var debugLevel = parseInt(process.env.NODE_DEBUG, 16);
 function debug () {
-  if (debugLevel & 0x2) sys.error.apply(this, arguments);
+  if (debugLevel & 0x2) util.error.apply(this, arguments);
 }
 var binding = process.binding('net');
 
@@ -532,7 +532,7 @@ function Stream (fd, type) {
     setImplmentationMethods(this);
   }
 };
-sys.inherits(Stream, stream.Stream);
+util.inherits(Stream, stream.Stream);
 exports.Stream = Stream;
 
 
@@ -783,7 +783,7 @@ Stream.prototype._writeOut = function (data, encoding, fd) {
   var leftOver = buffer.slice(off + bytesWritten, off + len);
   leftOver.used = leftOver.length; // used the whole thing...
 
-  //  sys.error('data.used = ' + data.used);
+  //  util.error('data.used = ' + data.used);
   //if (!this._writeQueue) initWriteStream(this);
 
   // data should be the next thing to write.
@@ -1093,7 +1093,7 @@ function Server (listener) {
     }
   };
 }
-sys.inherits(Server, events.EventEmitter);
+util.inherits(Server, events.EventEmitter);
 exports.Server = Server;
