+++ lib/repl.js
@@ -12,7 +12,7 @@
 
 // repl.start("node > ").context.foo = "stdin is fun";  // expose foo to repl context
 
-var sys = require('sys');
+var util = require('util');
 var Script = process.binding('evals').Script;
 var evalcx = Script.runInContext;
 var path = require("path");
@@ -41,7 +41,7 @@ function resetContext() {
 
 
 // Can overridden with custom print functions, such as `probe` or `eyes.js`
-exports.writer = sys.inspect;
+exports.writer = util.inspect;
 
 function REPLServer(prompt, stream) {
   var self = this;
@@ -59,7 +59,7 @@ function REPLServer(prompt, stream) {
   if (rli.enabled && !disableColors) {
     // Turn on ANSI coloring.
     exports.writer = function(obj, showHidden, depth) {
-      return sys.inspect(obj, showHidden, depth, true);
+      return util.inspect(obj, showHidden, depth, true);
     };
   }
 
@@ -167,9 +167,9 @@ REPLServer.prototype.readline = function (cmd) {
  *    (2) the leading text completed.
  *
  * Example:
- *  complete('var foo = sys.')
- *    -> [['sys.print', 'sys.debug', 'sys.log', 'sys.inspect', 'sys.pump'],
- *        'sys.' ]
+ *  complete('var foo = util.')
+ *    -> [['util.print', 'util.debug', 'util.log', 'util.inspect', 'util.pump'],
+ *        'util.' ]
  *
  * Warning: This eval's code like "foo.bar.baz", so it will run property
  * getter code.
@@ -253,7 +253,7 @@ REPLServer.prototype.complete = function (line) {
       // Intentionally excluding moved modules: posix, utils.
       var builtinLibs = ['assert', 'buffer', 'child_process', 'crypto', 'dgram',
         'dns', 'events', 'file', 'freelist', 'fs', 'http', 'net', 'path',
-        'querystring', 'readline', 'repl', 'string_decoder', 'sys', 'tcp', 'url'];
+        'querystring', 'readline', 'repl', 'string_decoder', 'util', 'tcp', 'url'];
       completionGroups.push(builtinLibs);
     }
   }
