+++ lib/stream.js
@@ -1,10 +1,10 @@
 var events = require('events');
-var inherits = require('sys').inherits;
+var util = require('util');
 
 function Stream () {
   events.EventEmitter.call(this);
 }
-inherits(Stream, events.EventEmitter);
+util.inherits(Stream, events.EventEmitter);
 exports.Stream = Stream;
 
 Stream.prototype.pipe = function (dest, options) {
