--- lib/utils.js
@@ -1,1 +1,0 @@
-sys.js
