+++ test/disabled/test-dns.js
@@ -2,8 +2,7 @@ common = require("../common");
 assert = common.assert;
 
 var dns = require("dns"),
-    child_process = require("child_process"),
-    sys = require("sys");
+    child_process = require("child_process");
 
 
 // Try resolution without callback
