+++ test/fixtures/print-10-lines.js
@@ -1,4 +1,3 @@
-puts = require('sys').puts;
 for (var i = 0; i < 10; i++) {
   console.log('count ' + i);
 }
