+++ test/simple/test-child-process-custom-fds.js
@@ -5,7 +5,6 @@ var assert = require('assert');
 var spawn  = require('child_process').spawn;
 var path   = require('path');
 var fs     = require('fs');
-var sys    = require('sys');
 
 function fixtPath(p) {
   return path.join(common.fixturesDir, p);
