+++ test/simple/test-child-process-exec-env.js
@@ -1,7 +1,6 @@
 common = require('../common');
 assert = common.assert;
-var exec = require('child_process').exec,
-    sys = require('sys');
+var exec = require('child_process').exec;
 success_count = 0;
 error_count = 0;
 response = "";
