+++ test/simple/test-crypto.js
@@ -9,7 +9,6 @@ try {
 }
 
 var fs = require('fs');
-var sys = require('sys');
 var path = require('path');
 
 // Test Certificates
