+++ test/simple/test-regression-object-prototype.js
@@ -1,5 +1,3 @@
-var sys = require('sys');
-
 //console.log('puts before');
 
 Object.prototype.xadsadsdasasdxx = function () {
