+++ tools/ronnjs/lib/ronn.js
@@ -4,7 +4,7 @@
  */
 
 var md = require(__dirname + '/ext/markdown');
-var sys = require('sys');
+var util = require('util');
 
 /* exports Ronn class
  * usage :
@@ -36,7 +36,7 @@ exports.Ronn = function(text, version, manual, date) {
 
 	function blockFilter(out, node, context) {
 		if (typeof node == "string") {
-			if (!node.match(/^\s*$/m)) sys.debug("unexpected text: " + node);
+			if (!node.match(/^\s*$/m)) util.debug("unexpected text: " + node);
 			return out;
 		}
 		var tag = node.shift();	
@@ -150,7 +150,7 @@ exports.Ronn = function(text, version, manual, date) {
 				out += "\n";
 			break;
 			default:
-				sys.debug("unrecognized block tag: " + tag);
+				util.debug("unrecognized block tag: " + tag);
 			break;
 		}
 		context.parent = fParent;
@@ -235,7 +235,7 @@ exports.Ronn = function(text, version, manual, date) {
 				}
 			break;
 			default:
-				sys.debug("unrecognized inline tag: " + tag);
+				util.debug("unrecognized inline tag: " + tag);
 			break;
 		}
 		context.parent = fParent;
